[SPCM95] Add WPT tests for payment credential creation
The credential creation is created for payments when the creation
option has a payment extension. For payment credentials, the
SPC implementation requires that authenticatorSelection has to
include:
{
  userVerification: 'required',
  residentKey: 'required',
  authenticatorAttachment: 'platform',
}

This CL adds the tests to verify this requirement. The tests do this
by verifying the "good" path and the "bad" path. The "bad" path is to
verify that setting any of the field to other values will lead to
"NotSupportedError". The "good" path is to verify that getting 3 fields
to be correct will lead to "NotAllowedError", because the test timeout
while the test is waiting for the user verification. It would be better
if the tests could bypass it, but the test has not been set up or not
being able to do it.

Bug: 1241090
Cq-Include-Trybots: luci.chromium.try:linux-wpt-identity-fyi-rel
Change-Id: I3d9aa0ea0e22a29bedf0034e3d57aefb81e9a2b6